requirements.txt
Django==5.0.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt5==5.15.10
asgiref==3.7.2
contourpy==1.2.1
crispy-bootstrap5==2024.2
crispy==0.7.4
cycler==0.12.1
dj-database-url==2.1.0
django-cookiebanner==0.2.8
django-crispy-forms==2.1
django-cryptography==1.1
django-environ==0.11.2
django-guardian==2.4.0
django-heroku==0.3.1
environ==1.0
fabio==2023.10.0
fonttools==4.51.0
gunicorn==21.2.0
h5py==3.11.0
hdf5plugin==4.4.0
kiwisolver==1.4.5
lxml==5.2.1
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pip~=24.1.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
pyparsing==3.1.2
pytest~=8.3.2
pytest-django~=4.8.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
silx==2.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
uuid~=1.30
whitenoise==6.6.0
xhtml2pdf~=0.2.16